Refactor schema package
