+++ ecs_model_deployer/src/lib/app.ts
@@ -15,10 +15,11 @@
 */
 
 import { AddPermissionBoundary } from '@cdklabs/cdk-enterprise-iac';
-import { Aspects, App } from 'aws-cdk-lib';
+import { App, Aspects } from 'aws-cdk-lib';
 import { LisaModelStack, LisaModelStackProps } from './lisa_model_stack';
 
-import { ConfigFile, ConfigSchema } from './ecs-schema';
+import { ConfigFile, ConfigSchema } from '#root/lib/schema';
+
 
 export const app = new App();
