+++ lib/api-base/ecsCluster.ts
@@ -48,9 +48,7 @@ import { StringParameter } from 'aws-cdk-lib/aws-ssm';
 import { Construct } from 'constructs';
 
 import { createCdkId } from '../core/utils';
-import { BaseProps } from '../schema';
-import { EcsSourceType } from '../cdk';
-import { Ec2Metadata, ECSConfig } from '../configSchema';
+import { BaseProps, EcsSourceType, Ec2Metadata, ECSConfig } from '#root/lib/schema';
 import { Vpc } from '../networking/vpc';
 
 /**
@@ -320,7 +318,7 @@ export class ECSCluster extends Construct {
         // ALB metric for ASG to use for auto scaling EC2 instances
         // TODO: Update this to step scaling for embedding models??
         const requestCountPerTargetMetric = new Metric({
-            metricName: ecsConfig.autoScalingConfig.metricConfig.AlbMetricName,
+            metricName: ecsConfig.autoScalingConfig.metricConfig.albMetricName,
             namespace: 'AWS/ApplicationELB',
             dimensionsMap: {
                 TargetGroup: targetGroup.targetGroupFullName,
