+++ lib/schema/cdk.ts
@@ -57,5 +57,4 @@ export enum AmiHardwareType {
     ARM = 'ARM64',
     NEURON = 'Neuron'
 }
-
 /* eslint-enable @typescript-eslint/no-duplicate-enum-values */
