+++ lib/user-interface/react/package.json
@@ -41,6 +41,7 @@
     "unraw": "^3.0.0"
   },
   "devDependencies": {
+    "@types/markdown-it": "^14.1.2",
     "@types/react": "^18.3.18",
     "@types/react-dom": "^18.3.5",
     "@types/redux-mock-store": "^1.5.0",
@@ -56,6 +57,8 @@
     "eslint-plugin-prettier": "^5.2.3",
     "eslint-plugin-react-hooks": "^4.6.2",
     "eslint-plugin-react-refresh": "^0.4.18",
+    "linkify-it": "^5.0.0",
+    "markdown-it": "^14.1.0",
     "postcss": "^8.5.1",
     "prettier": "^3.4.2",
     "redux-mock-store": "^1.5.5",
