+++ lib/user-interface/react/tailwind.config.mjs
