+++ lib/user-interface/react/tsconfig.json
@@ -1,12 +1,25 @@
 {
   "compilerOptions": {
+    "module": "ESNext",
+    "moduleResolution": "bundler",
     "target": "ES2020",
     "useDefineForClassFields": true,
     "lib": ["ES2020", "DOM", "DOM.Iterable"],
-    "module": "ESNext",
+    "baseUrl": ".",
+    "paths": {
+      "@/*": [
+        "./src/*"
+      ],
+      "#root/*": [
+        "../../../*"
+      ]
+    },
+    "types": [
+      "node",
+      "linkify-it"
+    ],
     "skipLibCheck": true,
     "checkJs": false,
-    "moduleResolution": "bundler",
     "allowImportingTsExtensions": true,
     "resolveJsonModule": true,
     "isolatedModules": true,
