+++ tsconfig.json
@@ -1,8 +1,20 @@
 {
   "compilerOptions": {
-    "target": "ES2020",
+    "moduleResolution": "node",
     "module": "commonjs",
-    "lib": ["es2020", "dom"],
+    "target": "ES2020",
+    "lib": ["ES2020", "dom"],
+    "types": [
+      "node",
+      "jest",
+      "linkify-it"
+    ],
+    "paths": {
+      "#root/*": [
+        "./*"
+      ]
+    },
+    "skipLibCheck": true,
     "resolveJsonModule": true,
     "esModuleInterop": true,
     "declaration": true,
@@ -21,6 +33,13 @@
     "strictPropertyInitialization": false,
     "typeRoots": ["./node_modules/@types"],
     "outDir": "dist/"
+  },
+  "ts-node": {
+    "esm": true,
+    "experimentalSpecifiers": true,
+    "nodeRegistration": {
+      "experimentalSpecifierResolution": "node"
+    }
   },
   "exclude": ["node_modules", "cdk.out", ".git", "dist", "lib/user-interface/react"]
 }
