+++ vector_store_deployer/package.json
@@ -4,7 +4,9 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "build": "tsc && cp package.json ./dist && cp ./src/cdk*json ./dist/ && cp ../VERSION ./dist && cd ./dist && npm i --omit dev",
+    "build": "tsc && npm run copy-deps && npm run install-run-deps",
+    "copy-deps": "cp package.json ./dist && cp ./src/cdk*json ./dist/ && cp ../VERSION ./dist",
+    "install-run-deps": "cd dist && npm i --omit dev --include-workspace-root",
     "clean": "rm -rf ./dist/",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
@@ -15,10 +17,12 @@
     "@cdklabs/cdk-enterprise-iac": "^0.0.512",
     "aws-cdk": "^2.153.0",
     "aws-cdk-lib": "^2.153.0",
-    "zod": "^3.23.8"
+    "zod": "^3.23.8",
+    "@types/node": "^22.4.1"
   },
   "devDependencies": {
     "@types/aws-lambda": "8.10.147",
-    "@types/node": "^22.4.1"
+    "@types/node": "^22.4.1",
+    "ts-node": "^10.9.2"
   }
 }
