+++ vector_store_deployer/src/index.ts
@@ -14,8 +14,8 @@
   limitations under the License.
 */
 
-import { spawnSync, spawn, ChildProcess } from 'child_process';
-import { readdirSync, symlinkSync, rmSync, mkdirSync } from 'fs';
+import { ChildProcess, spawn, spawnSync } from 'node:child_process';
+import { mkdirSync, readdirSync, rmSync, symlinkSync } from 'node:fs';
 
 /*
   cdk CLI always wants ./ to be writable in order to write cdk.context.json.
