+++ vector_store_deployer/src/lib/index.ts
@@ -16,8 +16,7 @@
 import { App } from 'aws-cdk-lib/core';
 import { OpenSearchVectorStoreStack } from './opensearch';
 import { PGVectorStoreStack } from './pgvector';
-import { RagRepositoryConfigSchema, RagRepositoryType } from '../../../lib/configSchema';
-import { PartialConfigSchema } from '../../../lib/schema';
+import { RagRepositoryConfigSchema, RagRepositoryType,PartialConfigSchema } from '#root/lib/schema';
 
 const app = new App();
