+++ vector_store_deployer/tsconfig.json
@@ -1,8 +1,23 @@
 {
   "compilerOptions": {
-    "target": "ES2020",
+    "lib": [
+      "es2020",
+      "dom"
+    ],
+    "types": [
+      "node"
+    ],
+    "paths": {
+      "@/*": [
+        "./src/*"
+      ],
+      "#root/*": [
+        "../*"
+      ]
+    },
+    "moduleResolution": "node",
     "module": "commonjs",
-    "lib": ["es2020", "dom"],
+    "target": "ES2020",
     "resolveJsonModule": true,
     "esModuleInterop": true,
     "declaration": true,
@@ -16,12 +31,15 @@
     "noImplicitReturns": true,
     "noFallthroughCasesInSwitch": false,
     "inlineSourceMap": true,
+    "skipLibCheck": true,
     "inlineSources": true,
     "experimentalDecorators": true,
     "strictPropertyInitialization": false,
-    "typeRoots": ["./node_modules/@types"],
+    "typeRoots": [
+      "../node_modules/@types"
+    ],
     "outDir": "dist"
   },
-  "include": ["src/**/*", "test/**/*", "../lib/core", "../lib/api-base", "../lib/utils", "../lib/schema.ts"],
+  "include": ["src/**/*", "test/**/*", "../lib/core", "../lib/api-base", "../lib/schema", "../lib/networking"],
   "exclude": ["node_modules", "dist", ".git"]
 }
